Fix kapt build issues and code quality
- Fix duplicate companion object in OcrProcessor (merged into one)
- Fix FeatureFlags import path
- Remove duplicate isInitialized function in ApiClient
- Update confidence type from Float to Double
- Fix test file type mismatch

+++ app/src/main/java/com/remitos/app/ui/screens/InboundViewModel.kt
@@ -152,12 +152,12 @@ class InboundViewModel(
         }
     }
 
-    fun updateOcrMetadata(text: String?, confidence: Map<String, Float>?) {
+    fun updateOcrMetadata(text: String?, confidence: Map<String, Double>?) {
         ocrTextBlob = text
         ocrConfidenceJson = confidenceToJson(confidence)
     }
 
-    private fun confidenceToJson(confidence: Map<String, Float>?): String? {
+    private fun confidenceToJson(confidence: Map<String, Double>?): String? {
         if (confidence.isNullOrEmpty()) return null
         return confidence.entries.joinToString(prefix = "{", postfix = "}") { (key, value) ->
             "\"$key\":$value"
